Added logic for selling items
But they need a rework. Need to change Data Structure and add a value to crops and give them a flag named Sell or Sellable. Can't work with groups here, since they get deleted when items get picked up and then are dropped again. Would need some extra logic to check, if a dropped item is sellable and instantiate them with the group again...